API completed on posts app

+++ posts/api/views.py
@@ -1,10 +1,12 @@
 from django.http import Http404
-from rest_framework.decorators import api_view
-from django.shortcuts import get_object_or_404
+from rest_framework.decorators import api_view,permission_classes
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
-from rest_framework.generics import UpdateAPIView
+from rest_framework.generics import ListAPIView
 from rest_framework.response import Response
 from rest_framework import status
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.pagination import PageNumberPagination
 from ..models import Post
 from .serializers import PostSerializer,PostUpdateSerializer
 from django.contrib.auth import models
@@ -12,16 +14,26 @@ from django.contrib.auth import models
 account = models.User.objects.get(username='bikrant')
 print(account)
 
-class PostList(APIView):
 
-    def get(self,request):
-        posts = Post.objects.all()
-        serializer = PostSerializer(posts,many=True)
+@api_view(['GET',])
+@permission_classes((IsAuthenticated,))
+def api_detail_view(request,slug):
+    try:
+        userpost = Post.objects.get(slug=slug)
+    except Post.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+    if request.method == "GET":
+        serializer = PostSerializer(userpost)
         return Response(serializer.data)
 
 
-    def post(self,request):
-        pass
+class PostList(ListAPIView):
+    queryset = Post.objects.all()
+    serializer_class = PostSerializer
+    authentication_classes = (TokenAuthentication,)
+    permission_classes = (IsAuthenticated,)
+    pagination_class = PageNumberPagination
+
 
 class PostUpdate(APIView):
 
@@ -38,11 +50,16 @@ class PostUpdate(APIView):
 
 
 @api_view(['PUT',])
+@permission_classes((IsAuthenticated,))
 def api_update_postview(request,pk):
     try:
         post = Post.objects.get(pk=pk)
     except Post.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+    requser = request.user
+
+    if post.user != requser:
+        return Response({'Response':"you dont have correct permissions to edit "})
 
     if request.method == "PUT":
         serializer = PostUpdateSerializer(post,data=request.data)
@@ -56,12 +73,16 @@ def api_update_postview(request,pk):
 
 
 @api_view(['DELETE',])
+@permission_classes((IsAuthenticated,))
 def api_delete_postview(request,pk):
     try:
         post = Post.objects.get(pk=pk)
     except Post.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
+    requser = request.user
+    if post.user != requser:
+        return Response({'Response': "you dont have correct permissions to delete "})
 
     if request.method == "DELETE":
         operation = post.delete()
@@ -74,8 +95,9 @@ def api_delete_postview(request,pk):
 
 
 @api_view(['POST',])
+@permission_classes((IsAuthenticated,))
 def api_create_postview(request):
-    account = models.User.objects.get(username='bikrant')
+    account = request.user
     post = Post(user=account)
     if request.method == "POST":
         serializer = PostUpdateSerializer(post,data=request.data)
